// client/src/app/core/generated/model/min-rps-play-result-dto.ts
/**
 * minFactory
 * minFactory OpenAPI specification. <br/> <a href=\"http://localhost:3000/openapi-json/\" target=\"_blank\">JSON specification</a>
 *
 * The version of the OpenAPI document: v0.0.0
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MinRpsMove } from './min-rps-move';
import { MinRpsResult } from './min-rps-result';


export interface MinRpsPlayResultDto { 
    player1Move: MinRpsMove;
    player2Move: MinRpsMove;
    result: MinRpsResult;
}
export namespace MinRpsPlayResultDto {
}